Revert: build single binary - build UI in static library
In order to build and load plugins and provide a single binary
build the UI in a static library that can be linked to the MainWindow and
let the plugins be able to build and be loaded by the mainwindow later.

This revert is mandatory in order to load plugins and
is necessary before updating plugin path and generating a package that
will install binary and plugins in some conventional place.