Load plugins from "plugins" directory relative to current directory
Look for plugins to load in the sub directory "plugins" relative
to the current directory. For this to work the software must be started from
where it is located. It is intended to run from a .desktop shortcut therefor
it will run properly.

This is a necessary step before adding build scripts to generate a package that
will install the software